backend/requirements.txt
altgraph==0.17.4
asgiref==3.9.1
certifi==2025.8.3
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.12.0
djangorestframework==3.15.2
idna==3.10
macholib==1.16.3
msgpack==1.1.1
packaging==25.0
redis==6.4.0
requests==2.32.3
setuptools==80.9.0
sqlparse==0.5.3
urllib3==2.5.0
